requirements.txt
alembic==1.4.3
annotated-types==0.6.0
appdirs==1.4.4
beautifulsoup4==4.7.1
blinker==1.7.0
bottle==0.12.24
bottle-websocket==0.2.9
cachetools==5.3.3
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.1
cssselect==1.2.0
Eel==0.11.1
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.9.0
Flask==3.0.3
future==0.18.3
gevent==22.10.2
gevent-websocket==0.10.1
google-ai-generativelanguage==0.6.1
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.0
googleapis-common-protos==1.63.0
greenlet==2.0.2
grpcio==1.62.1
grpcio-status==1.62.1
gTTS==2.3.1
httplib2==0.22.0
idna==3.4
imgurpython==1.1.7
importlib_metadata==7.1.0
itsdangerous==2.1.2
jieba3k==0.35.1
Jinja2==3.1.3
joblib==1.2.0
lxml==4.6.5
Mako==1.2.4
MarkupSafe==2.1.2
newspaper3k==0.2.8
nltk==3.8.1
pathvalidate==0.28.1
Pillow==9.0.1
praw==7.2.0
prawcore==2.0.0
proto-plus==1.23.0
protobuf==4.25.3
psaw==0.0.12
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pyparsing==3.1.2
python-dateutil==2.8.2
python-editor==1.0.4
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
requests-file==1.5.1
rsa==4.9
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.4
SQLAlchemy==1.3.23
tinysegmenter==0.3
tldextract==3.4.0
tqdm==4.64.1
twitter-text-python==1.1.1
typing_extensions==4.11.0
update-checker==0.18.0
uritemplate==4.1.1
urllib3==1.26.14
websocket-client==1.5.1
Werkzeug==3.0.2
whichcraft==0.6.1
youtube-dl==2021.12.17
zipp==3.18.1
zope.event==4.6
zope.interface==5.5.2
